Added heuristic for generating yields.
When generating yields, all reachable defs of a name are used.
This does not make sense in cases where an empty string is
assigned to a name (e.g., as a default to a parameter), and
a condition then checks to make sure the name isn't empty. In
that case, the analysis isn't strong enough to compare values
of variables to values checked in conditions. So, as a heuristic,
here we check to see if a) the name is included as a condition
for reaching this fragment, which means we are checking to see
if it is empty, and b) an empty string is given as a name
expansion. If this is the case, we remove this expansion.